Put push_solvers into a separate directory
